requirements.txt
Django==4.2
django-allauth==0.58.2
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.7
dj-database-url==2.1.0 